re-using AdminUrlResponse
